Made necessary changes to code
