+++ Weather Alert/Networking/SingleCityRequest.swift
@@ -12,8 +12,8 @@ import CoreData
 class SingleCityRequest: NetworkingRequest {
     
     init(cityId: Int) {
-        super.init()
-        path = "/weather?id=\(cityId)"
+        let path = "/weather?id=\(cityId)"
+        super.init(path: path)
     }
     
     func response(completionHandler: @escaping (City) -> Void) {
@@ -22,7 +22,7 @@ class SingleCityRequest: NetworkingRequest {
             let entity: NSEntityDescription? = NSEntityDescription.entity(forEntityName: "City", in: self.dataController.managedObjectContext)
             let object = City(entity: entity!, insertInto: self.dataController.managedObjectContext)
             
-            object.mapping(json: result!)
+            if let res = result { object.mapping(json: res) }
             completionHandler(object)
         }
     }
