+++ Weather Alert/Supporting Files/Global.swift
@@ -10,6 +10,6 @@ import Foundation
 
 struct AppShared {
     static var instances: AppShared = AppShared()
-    public let networking = Networking(config: URLSessionConfiguration.default, appId: "95fcc97498d7e406f5467cf3dc17f577")
+    public let networking = Networking(appId: "95fcc97498d7e406f5467cf3dc17f577")
     public let dataController = DataController()
 }
